Fix bug when generating a signed integer within a range
See https://github.com/smol-rs/fastrand/issues/15 ,
previously the result would not necessarily be within the range.

The issue is on line 271, because `len` may be negative.
For example, for the range -100 .. 40, we will perform `100 - -40` which
wraps around to -116.

From that negative number, we need to recover the correct distance
between 100 and 40. Previously, the code to do so was `-116 as u32`. But
that is incorrect since it wraps around the `u32::MAX` and gives
4294967180. Instead, what should be done is: `-116 as u8 as u32` to
give the correct distance: 140.

This is what this commit changes. However, it is not a one-line change
since the macro did not have access to the unsigned counterpart for each
signed integer type. There is thus one more parameter to rng_integer!
so that we can specify it manually.